adding newest python + selenium files
newest method to upload to lulu.com using selenium and python. pass.py
is being ignored as it stores the passwords. there are some example
pdfs in the in folder. --just tried this on a macintosh after
developing on a ubuntu machine. getting some errors and will need to
revisit/make more robust soon!